src/blog/requirements.txt
Flask
Flask-SQLAlchemy
pyjwt
psycopg2-binary
